requirements.txt
alembic==1.14.1
asyncpg==0.30.0
httpx==0.28.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
loguru==0.7.3
psycopg2-binary==2.9.6
pydantic-settings==2.7.1
pydantic==2.10.6
fastapi==0.115.8
slack_sdk==3.34.0
python-dotenv==1.0.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-multipart==0.0.20
authlib==1.3.0
itsdangerous==2.2.0
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
httpx==0.28.1
PyJWT==2.10.1
fastapi-users[sqlalchemy]==14.0.1
httpx-oauth==0.16.1
google-auth==2.27.0
